Run tests and check/lint source code in CI
Build with -Werror to not let any warnings through. Currently do not build
with -Werror when building for RPi since libsigc++ 2.10.0 headers does not
compile cleanly with GCC 8. Will enable -Werror for RPi build as well when
we have upgraded Yocto to Warrior that has libsigc++ 2.10.1 with a fix.
Could take in a temporary patch or workaround the problem with e.g.
CPLUS_INCLUDE_PATH, but probably not worth it.

tools/check is added for running clang-format and clang-tidy. Long term it
would be nice if there were targets generated by Meson. There already is a
clang-format target but need to add a target that does not modify source
code and returns correct exit code. There is an issue filed for generating
a clang-tidy target, see link in comment of tools/check.

A small format change has to be made to src/cli/main.cpp since formatting
changed between clang-format 6 (original version available in Ubuntu 18.04)
and clang-format 8 (version available in bionic-updates and installed in
tools/ci/Dockerfile in this commit).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>