Describe how to test tools/ci/Dockerfile
Passing the UID and GID args and the bind mount of the current working
directory in the same way as is done in Jenkins is perhaps not immediately
obvious, even if one has built Docker images before.

Main reason for adding the "name" and "work_dir" variables in the
description is to make it so the instructions do not have to change when
the template is used as a base for another project.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>